refactor:  함수, 변수, 클래스 이름 리팩토링 및 코드 수정
