+++ backend/src/main/java/org/cosmic/backend/domain/albumChat/api/AlbumChatAlbumLikeApi.java
@@ -1,19 +1,16 @@
-package org.cosmic.backend.domain.post.controller;
-
+package org.cosmic.backend.domain.albumChat.api;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.transaction.Transactional;
-import org.cosmic.backend.domain.post.dto.Comment.CommentDto;
-import org.cosmic.backend.domain.post.dto.Comment.CommentReq;
-import org.cosmic.backend.domain.post.dto.Comment.CreateCommentReq;
-import org.cosmic.backend.domain.post.dto.Comment.UpdateCommentReq;
-import org.cosmic.backend.domain.post.dto.Post.PostDto;
-import org.cosmic.backend.domain.post.service.CommentService;
-import org.cosmic.backend.globals.dto.ErrorResponse;
+import org.cosmic.backend.domain.albumChat.dto.albumChat.AlbumChatDto;
+import org.cosmic.backend.domain.albumChat.dto.albumlike.AlbumChatAlbumLikeDto;
+import org.cosmic.backend.domain.albumChat.dto.albumlike.AlbumChatAlbumLikeReq;
+import org.cosmic.backend.domain.albumChat.dto.albumlike.AlbumChatAlbumLikeResponse;
+import org.cosmic.backend.domain.albumChat.service.AlbumChatAlbumLikeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,10 +19,10 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/comment")
-public class CommentController {
+@RequestMapping("/api/albumchat/albumlike")
+public class AlbumChatAlbumLikeApi {//각 앨범의 총 좋아요
     @Autowired
-    private CommentService commentService;
+    private AlbumChatAlbumLikeService likeService;
 
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200",
@@ -36,7 +33,7 @@ public class CommentController {
                     }),
 
             @ApiResponse(responseCode = "404",
-                    description = "Not Found Post",
+                    description = "Not Found AlbumChat",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = ErrorResponse.class))
@@ -45,32 +42,8 @@ public class CommentController {
     }
     )
     @PostMapping("/give")
-    @Transactional
-    public List<CommentReq> getCommentsByPostId(@RequestBody PostDto post) {//postid에 있는 comment들을
-        return commentService.getCommentsByPostId(post.getPostId());
-    }
-    //어떤 user의 postid인지
-
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200",
-                    description = "Ok",
-                    content = {
-                            @Content(mediaType = "application/json",
-                                    schema = @Schema(implementation = String.class))
-                    }),
-            @ApiResponse(responseCode = "404",
-                    description = "Not Found User or Post",
-                    content = {
-                            @Content(mediaType = "application/json",
-                                    schema = @Schema(implementation = ErrorResponse.class))
-                    }
-            )
-    }
-    )
-    @PostMapping("create")
-    @Transactional
-    public CommentDto createComment(@RequestBody CreateCommentReq comment) {
-        return commentService.createComment(comment);
+    public List<AlbumChatAlbumLikeResponse> getAlbumChatAlbumLikeByAlbumChatId(@RequestBody AlbumChatDto albumChat) {//특정 post의 좋아요만 보게해주는
+        return likeService.getAlbumChatAlbumLikeByAlbumChatId(albumChat.getAlbumChatId());
     }
 
     @ApiResponses(value = {
@@ -80,15 +53,16 @@ public class CommentController {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = String.class))
                     }),
-            @ApiResponse(responseCode = "400",
-            description = "Not Match Post Or User",
-            content = {
-                    @Content(mediaType = "application/json",
-                            schema = @Schema(implementation = ErrorResponse.class))
-            }
-    ),
+
             @ApiResponse(responseCode = "404",
-                    description = "Not Found Post",
+                    description = "Not Found User or AlbumChat",
+                    content = {
+                            @Content(mediaType = "application/json",
+                                    schema = @Schema(implementation = ErrorResponse.class))
+                    }
+            ),
+            @ApiResponse(responseCode = "409",
+                    description = "Like Already Exists",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = ErrorResponse.class))
@@ -96,33 +70,30 @@ public class CommentController {
             )
     }
     )
-    @PostMapping("update")
-    @Transactional
-    public ResponseEntity<?> updateComment(@RequestBody UpdateCommentReq comment) {
-        commentService.updateComment(comment);
-        return ResponseEntity.ok("성공");
+    @PostMapping("/create")
+    public AlbumChatAlbumLikeReq albumChatCreateAlbumLike(@RequestBody AlbumChatAlbumLikeDto like) {
+        return likeService.albumChatCreateAlbumLike(like.getUserId(),like.getAlbumChatId());//유저 정보줌으로써 추가
     }
-
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200",
                     description = "Ok",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = String.class))
-                    }),
+                    }
+            ),
             @ApiResponse(responseCode = "404",
-                    description = "Not Found Comment",
+                    description = "Not Found Like",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = ErrorResponse.class))
                     }
             )
-    }
-    )
+    })
+
     @PostMapping("/delete")
-    @Transactional
-    public ResponseEntity<?> deleteComment(@RequestBody CommentDto commentdto) {
-        commentService.deleteComment(commentdto);
+    public ResponseEntity<?> deleteAlbumChatAlbumLike(@RequestBody AlbumChatAlbumLikeReq likedto) {
+        likeService.deleteAlbumChatAlbumLike(likedto.getAlbumChatAlbumLikeId());//해당 likeid를 제거
         return ResponseEntity.ok("성공");
     }
 }
