+++ backend/src/main/java/org/cosmic/backend/domain/auth/api/AuthApi.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.extern.log4j.Log4j2;
 import org.cosmic.backend.domain.auth.dto.UserLogin;
-import org.cosmic.backend.domain.user.service.UserService;
+import org.cosmic.backend.domain.user.applications.UserService;
 import org.cosmic.backend.globals.dto.ErrorResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
