+++ backend/src/main/java/org/cosmic/backend/domain/post/api/CommentApi.java
@@ -1,16 +1,19 @@
-package org.cosmic.backend.domain.albumChat.api;
+package org.cosmic.backend.domain.post.api;
+
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.cosmic.backend.domain.albumChat.dto.albumChat.AlbumChatDto;
-import org.cosmic.backend.domain.albumChat.dto.albumlike.AlbumChatAlbumLikeDto;
-import org.cosmic.backend.domain.albumChat.dto.albumlike.AlbumChatAlbumLikeReq;
-import org.cosmic.backend.domain.albumChat.dto.albumlike.AlbumChatAlbumLikeResponse;
-import org.cosmic.backend.domain.albumChat.service.AlbumChatAlbumLikeService;
+import jakarta.transaction.Transactional;
+import org.cosmic.backend.domain.post.dto.Comment.CommentDto;
+import org.cosmic.backend.domain.post.dto.Comment.CommentReq;
+import org.cosmic.backend.domain.post.dto.Comment.CreateCommentReq;
+import org.cosmic.backend.domain.post.dto.Comment.UpdateCommentReq;
+import org.cosmic.backend.domain.post.dto.Post.PostDto;
+import org.cosmic.backend.domain.post.service.CommentService;
+import org.cosmic.backend.globals.dto.ErrorResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,10 +22,10 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/albumchat/albumlike")
-public class AlbumChatAlbumLikeController {//각 앨범의 총 좋아요
+@RequestMapping("/api/comment")
+public class CommentApi {
     @Autowired
-    private AlbumChatAlbumLikeService likeService;
+    private CommentService commentService;
 
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200",
@@ -33,7 +36,7 @@ public class AlbumChatAlbumLikeController {//각 앨범의 총 좋아요
                     }),
 
             @ApiResponse(responseCode = "404",
-                    description = "Not Found AlbumChat",
+                    description = "Not Found Post",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = ErrorResponse.class))
@@ -42,8 +45,32 @@ public class AlbumChatAlbumLikeController {//각 앨범의 총 좋아요
     }
     )
     @PostMapping("/give")
-    public List<AlbumChatAlbumLikeResponse> getAlbumChatAlbumLikeByAlbumChatId(@RequestBody AlbumChatDto albumChat) {//특정 post의 좋아요만 보게해주는
-        return likeService.getAlbumChatAlbumLikeByAlbumChatId(albumChat.getAlbumChatId());
+    @Transactional
+    public List<CommentReq> getCommentsByPostId(@RequestBody PostDto post) {//postid에 있는 comment들을
+        return commentService.getCommentsByPostId(post.getPostId());
+    }
+    //어떤 user의 postid인지
+
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200",
+                    description = "Ok",
+                    content = {
+                            @Content(mediaType = "application/json",
+                                    schema = @Schema(implementation = String.class))
+                    }),
+            @ApiResponse(responseCode = "404",
+                    description = "Not Found User or Post",
+                    content = {
+                            @Content(mediaType = "application/json",
+                                    schema = @Schema(implementation = ErrorResponse.class))
+                    }
+            )
+    }
+    )
+    @PostMapping("create")
+    @Transactional
+    public CommentDto createComment(@RequestBody CreateCommentReq comment) {
+        return commentService.createComment(comment);
     }
 
     @ApiResponses(value = {
@@ -53,16 +80,15 @@ public class AlbumChatAlbumLikeController {//각 앨범의 총 좋아요
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = String.class))
                     }),
-
+            @ApiResponse(responseCode = "400",
+            description = "Not Match Post Or User",
+            content = {
+                    @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ErrorResponse.class))
+            }
+    ),
             @ApiResponse(responseCode = "404",
-                    description = "Not Found User or AlbumChat",
-                    content = {
-                            @Content(mediaType = "application/json",
-                                    schema = @Schema(implementation = ErrorResponse.class))
-                    }
-            ),
-            @ApiResponse(responseCode = "409",
-                    description = "Like Already Exists",
+                    description = "Not Found Post",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = ErrorResponse.class))
@@ -70,31 +96,33 @@ public class AlbumChatAlbumLikeController {//각 앨범의 총 좋아요
             )
     }
     )
-    @PostMapping("/create")
-    public AlbumChatAlbumLikeReq AlbumChatcreateAlbumLike(@RequestBody AlbumChatAlbumLikeDto like) {
-        System.out.println("*****"+like);
-        return likeService.AlbumChatcreateAlbumLike(like.getUserId(),like.getAlbumChatId());//유저 정보줌으로써 추가
+    @PostMapping("update")
+    @Transactional
+    public ResponseEntity<?> updateComment(@RequestBody UpdateCommentReq comment) {
+        commentService.updateComment(comment);
+        return ResponseEntity.ok("성공");
     }
+
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200",
                     description = "Ok",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = String.class))
-                    }
-            ),
+                    }),
             @ApiResponse(responseCode = "404",
-                    description = "Not Found Like",
+                    description = "Not Found Comment",
                     content = {
                             @Content(mediaType = "application/json",
                                     schema = @Schema(implementation = ErrorResponse.class))
                     }
             )
-    })
-
+    }
+    )
     @PostMapping("/delete")
-    public ResponseEntity<?> deleteAlbumChatAlbumLike(@RequestBody AlbumChatAlbumLikeReq likedto) {
-        likeService.deleteAlbumChatAlbumLike(likedto.getAlbumChatAlbumLikeId());//해당 likeid를 제거
+    @Transactional
+    public ResponseEntity<?> deleteComment(@RequestBody CommentDto commentdto) {
+        commentService.deleteComment(commentdto);
         return ResponseEntity.ok("성공");
     }
 }
