+++ backend/src/main/java/org/cosmic/backend/domain/post/service/CommentService.java
@@ -82,8 +82,6 @@ public class CommentService {
         }
         else {
             Comment comment1 = commentRepository.findByCommentId(comment.getCommentId());
-            System.out.println(comment1.getPost().getPostId());
-            System.out.println(comment.getPostId());
             if(comment1.getUser().getUserId()!=comment.getUserId())
             {
                 throw new NotMatchUserException();
